Fix bug, finally

+++ src/main/java/com/apap/tugas01/model/InstansiModel.java
@@ -1,23 +1,34 @@
 package com.apap.tugas01.model;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.List;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 @Entity
 @Table(name = "instansi")
 public class InstansiModel implements Serializable{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private BigInteger id;
 
     @NotNull
     @Size(max = 255)
@@ -46,13 +57,13 @@ public class InstansiModel implements Serializable{
         this.pegawaiList = pegawaiList;
     }
     
-    public long getId() {
-        return id;
-    }
+	public BigInteger getId() {
+		return id;
+	}
 
-    public void setId(long id) {
-        this.id = id;
-    }
+	public void setId(BigInteger id) {
+		this.id = id;
+	}
 
     public String getNama() {
         return nama;
