+++ src/main/java/com/apap/tugas01/repository/JabatanPegawaiDb.java
@@ -1,5 +1,6 @@
 package com.apap.tugas01.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.apap.tugas01.model.JabatanPegawaiModel;
 
 @Repository
-public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel,Long> {
+public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel, BigInteger> {
 	Optional<List<JabatanPegawaiModel>> findAllByPegawaiNip(String nip);
-	List<JabatanPegawaiModel> findAllByJabatanId(Long id);
+	List<JabatanPegawaiModel> findAllByJabatanId(BigInteger id);
 }
