+++ src/main/java/com/apap/tugas01/repository/ProvinsiDb.java
@@ -1,10 +1,12 @@
 package com.apap.tugas01.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.apap.tugas01.model.ProvinsiModel;
 
 @Repository
-public interface ProvinsiDb extends JpaRepository<ProvinsiModel,Long> {
+public interface ProvinsiDb extends JpaRepository<ProvinsiModel, BigInteger> {
 }
