+++ src/main/java/com/apap/tugas01/service/JabatanService.java
@@ -1,5 +1,6 @@
 package com.apap.tugas01.service;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,8 +9,8 @@ import com.apap.tugas01.model.JabatanModel;
 public interface JabatanService {
 	void addJabatan(JabatanModel jabatan);
 	void deleteJabatan(JabatanModel jabatan);
-	void updateJabatan(Long id, JabatanModel jabatan);
-	Optional<JabatanModel> findJabatanById(Long id);
+	void updateJabatan(BigInteger id, JabatanModel jabatan);
+	Optional<JabatanModel> findJabatanById(BigInteger id);
 	List<JabatanModel> viewAll();
 
 }
